+++ common/src/logging.rs
@@ -1,5 +1,6 @@
 use anyhow::Result;
 use clap::ValueEnum;
+use std::io::IsTerminal;
 use tracing_error::ErrorLayer;
 use tracing_subscriber::{Layer, layer::SubscriberExt, util::SubscriberInitExt};
 
@@ -13,7 +14,9 @@ pub enum DisplayMode {
 
 /// Initializes new logging to the console and returns worker guard that will flush logs on drop.
 pub fn initialize(app_name: &str) -> Result<tracing_appender::non_blocking::WorkerGuard> {
-    let (non_blocking_appender, guard) = tracing_appender::non_blocking(std::io::stdout());
+    let stdout = std::io::stdout();
+    let is_terminal = stdout.is_terminal();
+    let (non_blocking_appender, guard) = tracing_appender::non_blocking(stdout);
 
     let timer = time::format_description::parse("[year]-[month padding:zero]-[day padding:zero] [hour]:[minute]:[second]")?;
     let time_offset = time::UtcOffset::current_local_offset().unwrap_or(time::UtcOffset::UTC);
@@ -30,7 +33,7 @@ pub fn initialize(app_name: &str) -> Result<tracing_appender::non_blocking::Work
 
     let stdout = tracing_subscriber::fmt::layer()
         .with_timer(timer)
-        .with_ansi(true)
+        .with_ansi(is_terminal)
         .with_writer(non_blocking_appender)
         .with_filter(env_filter);