+++ receiver/src/utils.rs
@@ -1,17 +1,21 @@
 use common::{dc09::DC09Message, logging::DisplayMode};
 use std::borrow::Cow;
 
-pub fn build_response_message(msg: DC09Message, key: Option<&str>, nak: bool) -> String {
+use crate::server::AckMode;
+
+pub fn build_response_message(msg: DC09Message, key: Option<&str>, ack: AckMode) -> String {
     let was_encrypted = msg.was_encrypted();
-    let response = if nak {
-        DC09Message::nak()
-    } else {
-        DC09Message::ack(msg.account, msg.sequence)
+    let response = match ack {
+        AckMode::Ack => DC09Message::ack(msg.account, msg.sequence)
             .with_receiver(msg.receiver)
-            .with_line_prefix(msg.line_prefix)
+            .with_line_prefix(msg.line_prefix),
+        AckMode::Nak => DC09Message::nak(),
+        AckMode::Duh => DC09Message::duh(msg.account, msg.sequence)
+            .with_receiver(msg.receiver)
+            .with_line_prefix(msg.line_prefix),
     };
 
-    if was_encrypted && !nak {
+    if was_encrypted && ack == AckMode::Ack {
         if let Some(key) = key {
             response
                 .to_encrypted(key)