+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## 0.1.8 - 2026-02-05
+
+### Features
+
+- add optional `DUH` response for received messages
+- add `--show` switch for dialler and receiver
+
+### Bug fixes
+
+- disable ansi colors if output is not to a terminal
+
 ## 0.1.7 - 2025-10-15
 
 ### Features